gert

+++ T5.java
@@ -4,7 +4,7 @@
 public class T5 {
     public static void main(String[] args) {
         double sum = 0;
-        for (int index = 1; index <= 20; index++) {
+        for (double index = 1; index <= 20; index++) {
             double itemValue = 1;
             for (double number = index; number >= 1; number--) {
                 itemValue = itemValue * number;
